requirements.txt
asgiref==3.2.10
Django==1.11.7
gunicorn==20.0.4
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
sqlparse==0.3.1
